EDSC-4475: Converting AboutCSDAModal
